refactor(DTO): Moved the User and Group DTO to Kratos package
